backend/requirements.txt
flask
flask-cors
PyMuPDF
requests 
httpx
python-dotenv